move call code into common controller; add reconnect action
