+++ app/controllers/call/index.js
@@ -1,16 +1,12 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-  needs: ['application'],
+  needs: ['application', 'call'],
   actions: {
     next: function(){
-      console.log('init call model', this.get('model'));
-      var url = 'https://mayone-staging.herokuapp.com/calls';
-      var data = {
-        email: this.get('controllers.application.model.email'),
-        phone: this.get('phone')
-      };
-      Ember.$.post(url, data);
+      var appModel = this.get('controllers.application.model');
+      Ember.set(appModel, 'phone', this.get('phone'));
+      this.get('controllers.call').send('connect');
       this.transitionToRoute('call.script');
     }
   }
